Reconcile scoring logic with actual game constraints.
Tighten constraints substantially:
* Count() only accepts 1 or 2.
* SumCount() only accepts 5 combinations and count of 4.
* GreaterThan only accepts 5 combinations.

This is codified in a static set that can be checked with
Constraint.is_valid(). There is also a canonicalize() method to handle
the inevitable ordering issue in the checks.

Change the syntax for Count{1,2} and SumCount string parsing.

Add Constraint::from_list() to replace a couple of implementations that
parse such a list. Fail if the constraints are not unique because they
all are in the board game set.